requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.6
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0
